Buttons Group checkbox Input Control

Create buttons grouped that act like checkboxes.

checkboxGroupButtons(inputId, label = NULL, choices, selected = NULL,
  status = "default", size = "normal", direction = "horizontal",
  justified = FALSE, individual = FALSE, checkIcon = list())

Arguments

inputId
The input slot that will be used to access the value.

label
Input label.

choices
List of values to select from (if elements of the list are named then that name rather than the value is displayed to the user)

selected
The initially selected value.

status
Color of the buttons

size
Size of the buttons ('xs', 'sm', 'normal', 'lg')

direction
Horizontal or vertical.

justified
If TRUE, fill the width of the parent div.

individual
If TRUE, buttons are separated.

checkIcon
A list, if no empty must contain at least one element named 'yes' corresponding to an icon to display if the button is checked.

Value

A buttons group control that can be added to a UI definition.

Examples

## Not run: ------------------------------------
# ## Only run examples in interactive R sessions
# if (interactive()) {
# 
# ui <- fluidPage(
#   checkboxGroupButtons(inputId = "somevalue", choices = c("A", "B", "C")),
#   verbatimTextOutput("value")
# )
# server <- function(input, output) {
#   output$value <- renderText({ input$somevalue })
# }
# shinyApp(ui, server)
# }
## ---------------------------------------------