Search Input

A text input only triggered by hiting Enter or clicking search button

searchInput(inputId, label = NULL, value = "", placeholder = NULL,
  btnSearch = NULL, btnReset = NULL, width = NULL)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

value
Initial value.

placeholder
A character string giving the user a hint as to what can be entered into the control.

btnSearch
An icon for the button which validate the search.

btnReset
n icon for the button which reset the search.

width
The width of the input, e.g. '400px', or '100%'.

Examples

## Not run: ------------------------------------
# if (interactive()) {
#   ui <- fluidPage(
#     tags$h1("Search Input"),
#     br(),
#     searchInput(
#       inputId = "search", label = "Enter your text",
#       placeholder = "A placeholder",
#       btnSearch = icon("search"),
#       btnReset = icon("remove"),
#       width = "450px"
#     ),
#     br(),
#     verbatimTextOutput(outputId = "res")
#   )
# 
#   server <- function(input, output, session) {
#     output$res <- renderPrint({
#       input$search
#     })
#   }
# 
#   shinyApp(ui = ui, server = server)
# }
## ---------------------------------------------