setBackgroundImage {shinyWidgets}	R Documentation

Custom background image for your shinyapp

Description

Allow to change the background image of your shinyapp.

Usage

setBackgroundImage(src = NULL)

Arguments

src
Url or path to the image, if using local image, the file must be in www/ directory and the path not contain www/.

Examples

if (interactive()) {

library(shiny)
library(shinyWidgets)

ui <- fluidPage(
  tags$h2("Add a shiny app background image"),
  setBackgroundImage(src = "http://wallpics4k.com/wp-content/uploads/2014/07/470318.jpg")
)

server <- function(input, output, session) {

}

shinyApp(ui, server)

}

[Package shinyWidgets version 0.4.8 Index]