updateNoUiSliderInput {shinyWidgets}	R Documentation

Change the value of a no ui slider input on the client

Description

Change the value of a no ui slider input on the client

Usage

updateNoUiSliderInput(session, inputId, value = NULL, range = NULL,
  disable = FALSE)

Arguments

session
The session object passed to function given to shinyServer.

inputId
The id of the input object.

value
The new value.

range
The new range, must be of length 2 with c(min, max).

disable
logical, disable or not the slider, if disabled the user can no longer modify the slider value

Examples

## Not run: 

if (interactive()) {

 demoNoUiSlider("update")

}

## End(Not run)

[Package shinyWidgets version 0.4.8 Index]