Numeric Range Input

Create an input group of numeric inputs that function as a range input.

numericRangeInput(inputId, label, value, width = NULL, separator = " to ")

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

value
The initial value(s) for the range. A numeric vector of length one will be duplicated to represent the minimum and maximum of the range; a numeric vector of two or more will have its minimum and maximum set the minimum and maximum of the range.

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit().

separator
String to display between the start and end input boxes.

Examples

if (interactive()) {

### examples ----

# see ?demoNumericRange
demoNumericRange()

###  basic usage ----

library( shiny )
library( shinyWidgets )

ui <- fluidPage(

  tags$br(),

  numericRangeInput(
    inputId = "noui1", label = "Numeric Range Input:",
    value = c(100, 400)
  ),
  verbatimTextOutput(outputId = "res1")

)

server <- function(input, output, session) {

  output$res1 <- renderPrint(input$noui1)

}

shinyApp(ui, server)

}