searchInput {shinyWidgets}	R Documentation

Search Input

Description

A text input only triggered when Enter key is pressed or search button clicked

Usage

searchInput(inputId, label = NULL, value = "", placeholder = NULL,
  btnSearch = NULL, btnReset = NULL, resetValue = "", width = NULL)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

value
Initial value.

placeholder
A character string giving the user a hint as to what can be entered into the control.

btnSearch
An icon for the button which validate the search.

btnReset
An icon for the button which reset the search.

resetValue
Value used when reset button is clicked, default to "", if NULL value is not reset.

width
The width of the input, e.g. '400px', or '100%'.

Note

The two buttons ('search' and 'reset') act like actionButton, you can retrieve their value server-side with input$<INPUTID>_search and input$<INPUTID>_reset.

See Also

updateSearchInput to update value server-side.

Examples

## Not run: 
if (interactive()) {
  ui <- fluidPage(
    tags$h1("Search Input"),
    br(),
    searchInput(
      inputId = "search", label = "Enter your text",
      placeholder = "A placeholder",
      btnSearch = icon("search"),
      btnReset = icon("remove"),
      width = "450px"
    ),
    br(),
    verbatimTextOutput(outputId = "res")
  )

  server <- function(input, output, session) {
    output$res <- renderPrint({
      input$search
    })
  }

  shinyApp(ui = ui, server = server)
}

## End(Not run)

[Package shinyWidgets version 0.4.8 Index]