Change the value of a numeric range input

Change the value of a numeric range input

updateNumericRangeInput(session, inputId, label, value)

Arguments

session
The session object passed to function given to shinyServer.

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

value
The initial value(s) for the range. A numeric vector of length one will be duplicated to represent the minimum and maximum of the range; a numeric vector of two or more will have its minimum and maximum set the minimum and maximum of the range.