Change the value of a radio group buttons input on the client

Change the value of a radio group buttons input on the client

updateRadioGroupButtons(session, inputId, selected = NULL)

Arguments

session
The session object passed to function given to shinyServer.

inputId
The id of the input object.

selected
The value selected.